requirements.txt
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
google-auth==2.19.1
requests==2.31.0
urllib3==1.26.16